+++ R/treesWoffset.R
@@ -20,7 +20,7 @@
 #'
 #' # Do something with the converted trees - for instance, calculate the MCC summary tree
 #' # Then remove the dummy tip from the MCC tree
-#' mcc_tree = drop.dummy(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
+#' final_tree = drop.dummy.beast(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
 #'
 #' @export
 #' @importFrom stats median
@@ -84,7 +84,7 @@ offset.to.dummy.metadata = function(trees.file, log.file, output.file = NULL,
 #'
 #' # Do something with the converted trees - for instance, calculate the MCC summary tree
 #' # Then remove the dummy tip from the MCC tree
-#' mcc_tree = drop.dummy(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
+#' final_tree = drop.dummy.beast(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
 #'
 #' @export
 #' @importFrom stats median
@@ -107,7 +107,7 @@ offset.to.dummy = function(trees.file, log.file, output.file = NULL,
 }
 
 
-#' Remove dummy tip from tree, accounting for metadata on the tips
+#' Remove dummy tip from beast summary trees, accounting for metadata on the tips
 #'
 #' This method is designed to remove the dummy tip added on offset trees once postprocessing is
 #' complete (for instance once the summary tree has been built using TreeAnnotator).
@@ -122,10 +122,11 @@ offset.to.dummy = function(trees.file, log.file, output.file = NULL,
 #' @examples
 #' # Analyze the trees with dummy tips - for instance, calculate the MCC summary tree
 #' # Then remove the dummy tip from the MCC tree
-#' mcc_tree = drop.dummy(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
+#' final_tree = drop.dummy.beast(system.file("extdata", "ex_offset.MCC.tre", package = "EvoPhylo"))
 #'
 #' @export
-drop.dummy = function(tree.file, output.file = NULL, dummy.name = "dummy", convert.heights = TRUE) {
+#' @seealso [drop.dummy.mb()] (for the same function using summary trees with a "dummy" extant from Mr. Bayes)
+drop.dummy.beast = function(tree.file, output.file = NULL, dummy.name = "dummy", convert.heights = TRUE) {
   tmp = treeio::read.beast(tree.file)
 
   tip = which(tmp@phylo$tip.label == dummy.name)
@@ -147,6 +148,59 @@ drop.dummy = function(tree.file, output.file = NULL, dummy.name = "dummy", conve
     for(m in c("height_0.95_HPD", "height_range", "height_median", "height")) {
       tmp@data[[m]] = lapply(tmp@data[[m]], function(x) x - offset)
     }
+  }
+
+  if(!is.null(output.file)) write.beast.treedata(list(tmp), file = output.file)
+
+  list(tree = tmp, offset = offset)
+}
+
+
+
+#' Remove dummy tip from Mr. Bayes summary trees, accounting for metadata on the tips
+#'
+#' This method is designed to remove the dummy tip added to a dataset before running with Mr. Bayes.
+#'
+#' @param tree.file path to file containing the tree with dummy tip
+#' @param output.file path to file to write converted tree. If \code{NULL} (default), the tree is simply returned.
+#' @param dummy.name name of the added dummy tip, default \code{dummy}.
+#' @param convert.ages whether height metadata should be converted to height - offset (required to plot e.g. HPD intervals correctly). Default TRUE.
+#'
+#' @return list of \code{tree} converted tree (as treedata) ; and \code{offset} age of the youngest tip in the final tree
+#'
+#' @examples
+#' # Remove the dummy tip from the summary tree
+#' final_tree = drop.dummy.mb(system.file("extdata", "tree_mb_dummy.tre", package = "EvoPhylo"))
+#'
+#' @export
+#' @seealso [drop.dummy.beast()] (for the same function using summary trees with a "dummy" extant from BEAST2)
+drop.dummy.mb = function(tree.file, output.file = NULL, dummy.name = "dummy", convert.ages = TRUE) {
+
+  tmp = treeio::read.mrbayes(tree.file)
+
+  tmp@data$`prob+-sd` <- as.factor(tmp@data$`prob+-sd`)
+  tmp@data <- tmp@data %>% dplyr::mutate_if(is.character,as.numeric)
+
+  tip = which(tmp@phylo$tip.label == dummy.name)
+  tipn = which(tmp@data$node == as.character(tip))
+  node = tmp@phylo$edge[which(tmp@phylo$edge[,2] == tip),1]
+  noden = which(tmp@data$node == as.character(node))
+
+  offset = min(tmp@data$age_median[which(as.numeric(tmp@data$node) <= length(tmp@phylo$tip.label) &
+                                              as.numeric(tmp@data$node) != tip)])
+
+  tmp@phylo = ape::drop.tip(tmp@phylo, tip)
+  tmp@data = tmp@data[-c(tipn,noden),]
+  tmp@data$node = as.numeric(tmp@data$node)
+  tmp@data$node[(tmp@data$node) > tip] = tmp@data$node[(tmp@data$node) > tip] - 1
+  tmp@data$node[(tmp@data$node) > node] = tmp@data$node[(tmp@data$node) > node] - 1
+  tmp@data$node = as.character(tmp@data$node)
+
+  if(convert.ages) {
+    for(m in c("age_median", "age_mean")) {
+      tmp@data[[m]] = lapply(tmp@data[[m]], "-", offset)
+    }
+    tmp@data$age_0.95HPD <- lapply(tmp@data$age_0.95HPD,"-", offset)
   }
 
   if(!is.null(output.file)) write.beast.treedata(list(tmp), file = output.file)